Add files via upload
Firdge IOT is basically a project where a Raspberry pi connected with motion sensor and camera is attached inside a fridge. The latest photo is always updated in a database. Here we used a Dropbox API to store those photos. And This android app fetches those photos and shows in the display.